+++ src/operator.cpp
@@ -1,30 +1,30 @@
 #include "operator.h"
 
 
-Variable &operator+(Variable &lhs, int val) {
+Variable &operator+(Variable &lhs, double val) {
     Variable *rhs = new Variable(val);
     return lhs + (*rhs);
 }
 
-Variable &operator+(int val, Variable &rhs) {
+Variable &operator+(double val, Variable &rhs) {
     return rhs + val;
 }
 
-Variable &operator*(Variable &lhs, int val) {
+Variable &operator*(Variable &lhs, double val) {
     Variable *rhs = new Variable(val);
     return lhs * (*rhs);
 }
 
-Variable &operator*(int val, Variable &rhs) {
+Variable &operator*(double val, Variable &rhs) {
     return rhs * val;
 }
 
-Variable &operator-(Variable &lhs, int val) {
+Variable &operator-(Variable &lhs, double val) {
     Variable *rhs = new Variable(val);
     return lhs - (*rhs);
 }
 
-Variable &operator-(int val, Variable &rhs) {
+Variable &operator-(double val, Variable &rhs) {
     Variable *lhs = new Variable(val);
     return (*lhs) - rhs;
 }
